Alpha version of the Write API
This code is the client facing part of the new write API. It currently
validates all requests (in a JSON format for PUT/POST) against the GOCDB
schema. It then rejects any request that can't currently be done through
the API. The API currently supports changes to extension properties for
sites, services, and service endpoints. Appropriate HTTP error codes are
provided throughout.