Refactor deletion of service endpoint proeprties
The code for deleting service endpoint extension properties now more
closely matches that used elsewhere for deleteing extension properties.
This includes a change the the variables for the delete endpoint
properties function, so the unit tests and web portal have been updated
to reflect this (the portal functionality has been tested and the unit
tests pass).