Stop endpoint urls needing to start with [a-z]+://
Alter the checks on endpoint URLs so they no longer require protocol://
at the start. This originally happned in two places. The view had some
javascript doing a check with a quite involved regex, this now
optionally allows 'something://'. the PHP function filter_var was also
used within the edit and add endpoint functions in the service service,
this has now been removed as it requires some sort  of protocol. There
is still a check against the schema, which is much more lenient.