Bug fixes in unit tests
These were found after running the unit tests against MySQL for the
first time for a while. The following changes have been made:
* Moved getId commnad to after flush in sampleFixtureDataa1, the id can
not be gurenteed to exist until the flush has been performed.
* Add flush statments to the ExtensionTest unit test. These flush the
admin user after creation, ensuring that the user is in the DB for the
RoleAuthorisationService to find.
* Remove old commented out code from ExtensionsTest.php