Add fields to GOCDB xml schema for boolians
This commit adds new fields to the gocdb XML Schema for service and
endpoint entities representing boolean properties these enteties have
but were not previously in the schema. They do not have a length or
regex defined, which means that the validate service will now allow any
value for those fields, but it will also recognise those fields are part
of that entitiy. An update to the validate service will follow, but this
commit is functionally complete.