Changes to exception handling in write API
A funciton is created that sets the http response code and then throws
an exception. This replaces over 40 instances where a similar code pattern
was used. This is an incremental improvement - there are however better
ways of handling exceptions (for instance extending the exception class
to include a resposne code).