Specify matrix values in job names
Job names automatically include matrix values, but it is not always clear what they correspond to. Instead, they can be explicitly passed to the name. This also avoids unnecessarily including "experimental". The version of SQLite depends on the current environment, as it is pre-installed, so we point to the documentation instead of including a version in the name.