Add functions to add identifiers to users
These functions allow identifiers to be added to users as key/value pairs, containing the authentication type and associated ID string. The functions are based on similar functions used to add extension properties to sites and services, and will require revisiting (e.g. validation). When adding the first identifier, certificateDn is currently overwritten by user ID, so the value is non-null and unique. An alternative placeholder may be used later.