Update controller and view for admin editing user identifiers
Updated functionality that allowed admins to edit user DNs to update user identifiers, including updating the ID string and/or auth type for an identifier, except the one they are currently using. If a user does not have identifiers, this also allows one to be added, if it matches their current certificateDn.