Add toggle link for scope help information
Similar to the downtimes calendar view, this hides the scope help information by default, but allows it to be viewed by clicking a link. This means the list of scopes is displayed first, but it is not necessary to scroll past the entire table to see the help information. The link text also changes on clicking.